refactor(user): 优化用户实体与DTO转换逻辑
- 使用MapStruct实现UserDO与UserInfo/BasicUserInfo的双向转换
- 添加Long与String类型ID的转换方法，支持前后端数据转换
- 更新UserInfo增加购买资格判断方法userCanBuy
- 引入lombok.Setter简化UserDO代码
- 修改UserMapper为MyBatis注解映射，提供基础CRUD能力
- 调整pom依赖，引入MapStruct及lombok-mapstruct-binding支持
- 移除.idea目录配置文件，清理无用项目配置
- 修正测试用例中lambda表达式和循环变量类型错误
- 微调CommonUtilsTest中集合转换测试逻辑
- 优化RestClientUtils日志引入方式，替代Sl4j注解
- 统一项目Java版本为21，完善maven-compiler-plugin配置

+++ finance-common/src/test/java/cn/com/mz/app/finance/common/utils/IDUtilsTest.java
@@ -261,7 +261,7 @@ class IDUtilsTest {
     void testMaxDbIndex() {
         // 测试能否产生最大库索引（15）
         boolean foundMaxDbIndex = false;
-        for (int i = 13000000000; i < 19999999999L; i += 100000000) {
+        for (long i = 13000000000L; i < 19999999999L; i += 100000000L) {
             try {
                 String phone = String.valueOf(i);
                 if (phone.length() == 11 && phone.matches("^1[3-9]\\d{9}$")) {
@@ -285,7 +285,7 @@ class IDUtilsTest {
     void testMaxTableIndex() {
         // 测试能否产生最大表索引（511）
         boolean foundMaxTableIndex = false;
-        for (int i = 13000000000; i < 19999999999L; i += 100000000) {
+        for (long i = 13000000000L; i < 19999999999L; i += 100000000L) {
             try {
                 String phone = String.valueOf(i);
                 if (phone.length() == 11 && phone.matches("^1[3-9]\\d{9}$")) {
